# perf: Don't copy contents of ast.Text